refactor: merge handler_body to user_body
