--- internal/handlers/dto/handler_body.go
@@ -1,14 +1,0 @@
-package dto
-
-type ResetPasswordCreateRequestBody struct {
-	Email string `json:"email" validate:"required,email"`
-}
-
-type ResetPasswordRequestBody struct {
-	Token    string `json:"token" validate:"required"`
-	Password string `json:"password" validate:"required"`
-}
-
-type VerifyRequestBody struct {
-	Token string `json:"token" validate:"required"`
-}
